# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  imageUrl: string;
  wholesalePrice: number;
  suggestedPrice: number;
  affiliateMargin: number;
  description: string;
  deliveryCost: number;
  inStock: boolean;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  NOUVELLE: 'NOUVELLE',
  CONFIRMEE: 'CONFIRMEE',
  EN_COURS_LIVRAISON: 'EN_COURS_LIVRAISON',
  LIVREE: 'LIVREE',
  RETOURNEE: 'RETOURNEE',
  ANNULEE: 'ANNULEE',
} as const;

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  customerFirstName: string;
  customerLastName: string;
  customerPhone: string;
  city: string;
  fullAddress: string;
  salePriceAffiliate: number;
  wholesalePrice: number;
  deliveryCost: number;
  netMargin: number;
  status: OrderStatus;
  /** @nullable */
  deliveryNote?: string | null;
  createdAt: string;
}

export interface OrderInput {
  productId: number;
  customerFirstName: string;
  customerLastName: string;
  customerPhone: string;
  city: string;
  fullAddress: string;
  salePriceAffiliate: number;
  deliveryNote?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  NOUVELLE: 'NOUVELLE',
  CONFIRMEE: 'CONFIRMEE',
  EN_COURS_LIVRAISON: 'EN_COURS_LIVRAISON',
  LIVREE: 'LIVREE',
  RETOURNEE: 'RETOURNEE',
  ANNULEE: 'ANNULEE',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface WalletBalance {
  withdrawableBalance: number;
  pendingBalance: number;
  totalEarned: number;
  minimumWithdrawal: number;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  EN_TRAITEMENT: 'EN_TRAITEMENT',
  PAYE: 'PAYE',
} as const;

export interface Withdrawal {
  id: number;
  amount: number;
  status: WithdrawalStatus;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  ribNumber?: string | null;
  requestedAt: string;
  /** @nullable */
  paidAt?: string | null;
}

export interface WithdrawalInput {
  amount: number;
}

export interface AnalyticsSummary {
  deliveryRate: number;
  returnRate: number;
  totalDelivered: number;
  totalReturned: number;
  totalShipped: number;
}

export interface DailyProfit {
  date: string;
  profit: number;
  orders: number;
}

export interface CityRevenue {
  city: string;
  revenue: number;
  orders: number;
}

export interface ProductStat {
  productId: number;
  productName: string;
  sales: number;
  revenue: number;
}

export interface DashboardStats {
  withdrawableBalance: number;
  pendingBalance: number;
  totalEarned: number;
  totalOrders: number;
  newOrdersToday: number;
}

export interface Profile {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  city: string;
  brandName: string;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  ribNumber?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
}

export interface ProfileUpdate {
  fullName?: string;
  phone?: string;
  email?: string;
  city?: string;
  brandName?: string;
  bankName?: string;
  ribNumber?: string;
  paymentMethod?: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
};

export type ListOrdersParams = {
status?: string;
search?: string;
};