# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  deliveryRate: number;
  returnRate: number;
  totalDelivered: number;
  totalReturned: number;
  totalShipped: number;
}