# lib/api-zod/src/generated/types/cityRevenue.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CityRevenue {
  city: string;
  revenue: number;
  orders: number;
}