# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  customerFirstName: string;
  customerLastName: string;
  customerPhone: string;
  city: string;
  fullAddress: string;
  salePriceAffiliate: number;
  wholesalePrice: number;
  deliveryCost: number;
  netMargin: number;
  status: OrderStatus;
  /** @nullable */
  deliveryNote?: string | null;
  createdAt: string;
}