# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  productId: number;
  customerFirstName: string;
  customerLastName: string;
  customerPhone: string;
  city: string;
  fullAddress: string;
  salePriceAffiliate: number;
  deliveryNote?: string;
}