# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  imageUrl: string;
  wholesalePrice: number;
  suggestedPrice: number;
  affiliateMargin: number;
  description: string;
  deliveryCost: number;
  inStock: boolean;
}