# lib/api-zod/src/generated/types/productStat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductStat {
  productId: number;
  productName: string;
  sales: number;
  revenue: number;
}