# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  city: string;
  brandName: string;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  ribNumber?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
}