# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Digital Ecom Land - COD Affiliation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  fullName?: string;
  phone?: string;
  email?: string;
  city?: string;
  brandName?: string;
  bankName?: string;
  ribNumber?: string;
  paymentMethod?: string;
}